refactor(serializers): move one directory up
It should be a sibling of the protocols, not a child.
