+++ trame_vtk/modules/vtk/serializers/helpers.py
