+++ trame_vtk/modules/vtk/serializers/lights.py
