+++ trame_vtk/modules/vtk/serializers/properties.py
