+++ trame_vtk/modules/vtk/serializers/textures.py
