+++ trame_vtk/modules/vtk/serializers/utils.py
